Update legacy-amd.md
Note that the HD 5550 is also supported.

My own card has a device ID of `68da` which is not present in the macOS drivers and requires a spoof to `68d9`, with which it works perfectly fine. Besides HS, I have tested it in Monterey with OCLP and a mostly passing result.